+++ lite/arm/math/sequence_pool.h
@@ -12,8 +12,11 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include <vector>
 #pragma once
+#include <vector>
+#include "lite/core/kernel.h"
+#include "lite/core/op_registry.h"
+#include "lite/core/tensor.h"
 
 namespace paddle {
 namespace lite {
@@ -24,43 +27,43 @@ template <typename T>
 void seq_pool_sum(const T* din,
                   T* dout,
                   const std::vector<uint64_t> lod,
-                  int64_t width);
+                  int width);
 
 template <typename T>
 void seq_pool_average(const T* din,
                       T* dout,
                       const std::vector<uint64_t> lod,
-                      int64_t width);
+                      int width);
 
 template <typename T>
 void seq_pool_sqrt(const T* din,
                    T* dout,
                    const std::vector<uint64_t> lod,
-                   int64_t width);
+                   int width);
 
 template <typename T>
 void seq_pool_max(const T* din,
                   T* dout,
                   const std::vector<uint64_t> lod,
-                  int64_t width);
+                  int width);
 
 template <typename T>
 void seq_pool_min(const T* din,
                   T* dout,
                   const std::vector<uint64_t> lod,
-                  int64_t width);
+                  int width);
 
 template <typename T>
 void seq_pool_first(const T* din,
                     T* dout,
                     const std::vector<uint64_t> lod,
-                    int64_t width);
+                    int width);
 
 template <typename T>
 void seq_pool_last(const T* din,
                    T* dout,
                    const std::vector<uint64_t> lod,
-                   int64_t width);
+                   int width);
 
 }  // namespace math
 }  // namespace arm
